moved interactive shell to its own package
