+++ internal/shell/shell_cmd_test.go
@@ -1,4 +1,4 @@
-package main
+package shell
 
 import (
 	"fmt"
