+++ internal/shell/shell_wasm.go
@@ -1,0 +1,7 @@
+//+build wasm
+
+package shell
+
+func Shell() error {
+	panic("not implemented")
+}
